pyproject.toml
[project]
name = "alfred"
version = "2.0.0"
description = "Alfred - A modern LangGraph-based multi-agent assistant for kitchen, fitness, and wine management"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]

dependencies = [
    # LLM Orchestration
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langsmith>=0.1.0",
    
    # Structured Outputs
    "instructor>=1.4.0",
    "openai>=1.50.0",
    
    # Database
    "supabase>=2.0.0",
    
    # Validation & Settings
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    
    # Utilities
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",

    # Recipe Import
    "recipe-scrapers>=15.0.0",
    "extruct>=0.17.0",
    
    # Web UI
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.12",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
alfred = "alfred_kitchen.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alfred", "src/alfred_kitchen", "src/onboarding"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

